// app/utils/sanity/types/sanity.types.ts
export type Twitter = {
  _type: 'twitter'
  url?: string
}
// ...
export type ImageWithMetadata = {
  _type: 'imageWithMetadata'
  asset?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
  }
  hotspot?: SanityImageHotspot
  crop?: SanityImageCrop
  src?: string
  caption?: string
  maxWidth?: number
  alt?: string
}

export type DescriptionText = Array<{
  children?: Array<{
    marks?: Array<string>
    text?: string
    _type: 'span'
    _key: string
  }>
  style?: 'normal' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6' | 'blockquote'
  listItem?: 'bullet' | 'number'
  markDefs?: Array<{
    href?: string
    _type: 'link'
    _key: string
  }>
  level?: number
  _type: 'block'
  _key: string
}>

export type CodeSandbox = {
  _type: 'codeSandbox'
  url?: string
}

export type CodePen = {
  _type: 'codePen'
  url?: string
}
// ...
export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export type Code = {
  _type: 'code'
  language?: string
  filename?: string
  code?: string
  highlightedLines?: Array<number>
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol